requirements.txt
# EduParser Dependencies
streamlit>=1.28.0
groq>=0.4.0
pandas>=2.0.0
openpyxl>=3.1.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
